requirements.txt
flask==3.1.0
requests==2.32.3
discord.py==2.4.0
python-dotenv==1.0.1
